+++ docs/_logo.html
@@ -43,6 +43,7 @@ nav#TOC {
 .quarto-title .title {
   margin-left: 180px;
   text-align: center;
+  color: white;
 }
 </style>
 
@@ -53,8 +54,8 @@ nav#TOC {
 <div class="logo-container">
   <img src="MIR.png" alt="Mission Iconic Reef Logo">
   <img src="noaa_logo.png" alt="NOAA Logo">
-</div>
 
+</div>
 <script>
 document.addEventListener("DOMContentLoaded", function() {
   // Select all anchor tags
@@ -71,3 +72,16 @@ document.addEventListener("DOMContentLoaded", function() {
   });
 });
 </script>
+
+<div style="text-align:left; margin-top: -1.5em;">
+  <p style="font-size:16px; margin-bottom:0.2em;">
+    <a href="https://github.com/jeremiaheb">Jeremiah Blondeau</a><sup>1</sup>,
+    Jay Grove<sup>1</sup>,
+    Rob Harper<sup>1,2</sup>
+  </p>
+
+  <p style="font-size:14px; margin-top:0.2em;">
+    1 NOAA National Marine Fisheries Service, Southeast Fisheries Science Center<br>
+    2 University of Miami, Cooperative Institute for Marine and Atmospheric Studies
+  </p>
+</div>
