+++ docs/references.bib
@@ -47,3 +47,37 @@
   url         = {https://doi.org/10.36967/2310167}
 }
 
+@article{Smith2011,
+  author        = {Steven G. Smith and Jerald S. Ault and James A. Bohnsack and Douglas E. Harper and Jiangang Luo and David B. McClellan},
+  title         = {Multispecies survey design for assessing reef‐fish stocks, spatially explicit management performance, and ecosystem condition},
+  journal       = {Fisheries Research},
+  volume        = {109},
+  number        = {1},
+  pages         = {25--41},
+  year          = {2011},
+  doi           = {10.1016/j.fishres.2011.01.012},
+  issn          = {0165-7836}
+}
+
+@misc{Grove2021,
+  author       = {Laura Jay W. Grove and Jeremiah Blondeau and Jerald S. Ault},
+  title        = {National Coral Reef Monitoring Program’s reef fish visual census metadata for the U.S. Caribbean},
+  year         = {2021},
+  howpublished = {NOAA Southeast Fisheries Science Center},
+  doi          = {10.25923/39r6-9578},
+  url          = {https://repository.library.noaa.gov/view/noaa/67188},
+  note         = {CC0 Public Domain}
+}
+
+@article{Bryan2016,
+  author        = {David R. Bryan and Steven G. Smith and Jerald S. Ault and Michael W. Feeley and Charles W. Menza},
+  title         = {Feasibility of a Regionwide Probability Survey for Coral Reef Fish in Puerto Rico and the U.S. Virgin Islands},
+  journal       = {Marine and Coastal Fisheries},
+  volume        = {8},
+  number        = {1},
+  pages         = {135--146},
+  year          = {2016},
+  doi           = {10.1080/19425120.2015.1082520},
+  url           = {https://doi.org/10.1080/19425120.2015.1082520},
+  issn          = {1942-5120}
+}
